chore(deps): bump CI and Docker Go pins
