+++ .github/workflows/release.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Set up Go
         uses: actions/setup-go@v6
         with:
-          go-version: stable
+          go-version: 1.26.1
           cache: true
 
       - name: Run GoReleaser
